--- a/src/vocoder/mlsa.rs
+++ b/src/vocoder/mlsa.rs
@@ -68,7 +68,7 @@ impl MelLogSpectrumApproximation {
     fn df2(&mut self, x: &mut f64, alpha: f64, coefficients: &'_ Coefficients) {
         let mut out = 0.0;
         for i in (1..=self.pd).rev() {
-            self.d22[i] = self.fir(self.d22[i - 1], alpha, coefficients, i - 1);
+            self.d22[i] = Self::fir(&mut self.d21[i - 1], self.d22[i - 1], alpha, coefficients);
             let v = self.d22[i] * self.ppade[i];
             *x += if i & 1 != 0 { v } else { -v };
             out += v;
@@ -77,9 +77,8 @@ impl MelLogSpectrumApproximation {
         *x += out;
     }
 
-    fn fir(&mut self, x: f64, alpha: f64, coefficients: &'_ Coefficients, i: usize) -> f64 {
+    fn fir(d: &mut [f64], x: f64, alpha: f64, coefficients: &'_ Coefficients) -> f64 {
         let aa = 1.0 - alpha * alpha;
-        let d = &mut self.d21[i];
         d[0] = x;
         d[1] = aa * d[0] + alpha * d[1];