main
```
test bonsai        ... bench:  46,765,787 ns/iter (+/- 129,285)
test bonsai_letter ... bench: 127,007,150 ns/iter (+/- 188,110)
test is_bonsai     ... bench:  71,291,024 ns/iter (+/- 3,036,153)
```
8728a07
```
test bonsai        ... bench:  20,215,567 ns/iter (+/- 64,065)
test bonsai_letter ... bench:  56,022,135 ns/iter (+/- 248,099)
test is_bonsai     ... bench:  31,164,132 ns/iter (+/- 178,681)
```
af373f3
```
test bonsai        ... bench:  19,108,283 ns/iter (+/- 97,274)
test bonsai_letter ... bench:  53,329,146 ns/iter (+/- 226,383)
test is_bonsai     ... bench:  29,430,791 ns/iter (+/- 53,959)
```